User reward points done

+++ buspass-ws-master/src/main/java/com/sctrcd/buspassws/model/BuyerProfile.java
@@ -31,8 +31,8 @@ public class BuyerProfile {
 		return rewardPoints;
 	}
 
-	public void setRewardPoints(int rewardPoints) {
-		this.rewardPoints = rewardPoints;
+	public void setRewardPoints(double rewardPoints) {
+		this.rewardPoints = (int)rewardPoints;
 	}
 
 	public BuyerCategory getCategory() {
